feat(create-pull-request.yml): add workflow to automatically create a pull request when a branch is created
The new file `.github/workflows/create-pull-request.yml` contains a
workflow that is triggered when a branch is created. It runs on the
`ubuntu-latest` environment and skips execution if the actor is
`dependabot[bot]`. The workflow performs the following steps:
1. Checks out the repository.
2. Retrieves the name of the created branch.
3. Creates a pull request using the `repo-sync/pull-request@v2` action, targeting the `develop` branch and using the branch name as the pull request title.
This workflow automates the process of creating pull requests when new branches are created, improving collaboration and code review processes.