feat(lint.yaml): add linting workflow for golang code
A new lint.yaml file is added to the .github/workflows directory. This
file contains a workflow that runs a linter on pull requests targeting
branches other than master. The workflow runs on an Ubuntu 20.04 runner
and performs the following steps:
1. Checks out the code.
2. Sets up the Go environment with version 1.20.4.
3. Runs the golangci-lint action with version 1.52 to perform linting on the code.
The linting workflow is added to ensure code quality and adherence to best practices in the golang codebase.