chore(.kodiak.toml): add .kodiak.toml configuration file for Kodiak CI
The .kodiak.toml file is added to the repository to configure the
behavior of the Kodiak CI tool. This configuration specifies that PRs
opened by the "dependabot" user and containing "minor" or "patch"
version upgrades should be automatically merged. "major" version
upgrades will be ignored. Additionally, the configuration specifies that
the "dependabot" user should be ignored when checking for stale
dependency upgrades.